[Internal] Tests: Add contract test owners (#1668)
* Reordering tests

* owners
